refactor: 사용하지 않는 Reservations 삭제 및 테스트 추가
